SW474357: Remove not working Server Identify LED
The IndicatorLED at /redfish/v1/Systems/system does not work and
has never worked.

The problem is it calls xyz.openbmc_project.LED.Controller.identify
which IBM does not use.

https://github.com/openbmc/bmcweb/blob/944ffaf9ccae2a084cd650370facfb3f76701186/redfish-core/lib/systems.hpp#L650

With OP940 closing in and this not being a needed Redfish Property,
remove for now.

https://github.com/ibm-openbmc/dev/issues/1202 tracks fixing
this. Will leave the defect open but defer out of OP940 if
this workaround is accepted.

Tested: Loaded bmcweb on a Witherspoon. System Resource looks
        good as do others.
Signed-off-by: Gunnar Mills <gmills@us.ibm.com>