Return invalid path in REST error response
To be backwards compatible with OP910-OP930, need to return the invalid
path in the error message on REST api's. The xcat team relies on this
path to do post processing work.

Verified with xcat team that it is only the "path or object not found:"
error that requires this.

Tested:

Previous Response:
{
  "data": {
    "description": "org.freedesktop.DBus.Error.FileNotFound: path or object not found"
  },
  "message": "404 Not Found",
  "status": "error"
}

New Response:
{
  "data": {
    "description": "org.freedesktop.DBus.Error.FileNotFound: path or object not found: /xyz/openbmc_project/sensors/chassis/PowerSupplyRedundancy"
  },
  "message": "404 Not Found",
  "status": "error"
}

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>