better password change failure message
This gives a better error message when the password fails to change.
specifically, Linux PAM function pam_chauthtok which returns
PAM_AUTHTOK_ERR for any number of reasons, including the common case
of the new password not meeting complexity requirements.
This patchset articulates that error code and gives better messages
to explain what is going on.

See upsteam work here:
https://redfishforum.com/thread/246/message-send-patch-password-failure

This changes behavior of the Redfish API
PATCH /redfish/v1/AccountService/Accounts/{account}
with a new Password property to give HTTP status 400 bad request
(was 500 server error) with the existing Redfish AccountNotModified
message, and include a new IBM OEM message to further explain.

The handler for /xyz/openbmc_project/user/root/action/SetPassword
(the legacy route) was removed because it was dead code.  Note that
the SetPassword API was added as a special case in OP940 under the
/xyz/<path> route.  This handler works as before.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>