default isConfigureSelfOnly for saved session
This ensures the default value for the UserSession.isConfigureSelfOnly
field is false, in case a newer version of BMCWeb reads session data
that was written by an older version of BMCWeb.

Before this fix, it is possible that a user could login, upgrade the
BMC to a version that has support for logging in whern your password
is expired, and when they continue to use the same session as before,
BMCWeb would act as if their password was expired.
The recovery would be to login again.

Tested: yes; users with good password still have good passwords after
bmcweb restart.  Users with ConfigureSelfOnly continue to have that
limitation after bmcweb restart.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>