bmc dump: fix upstream rebase issue
This fixes SW519308

Upstream changed the dump interfaces and when that changes was rebased
into downstream a part of that patch set was missed.

https://github.com/openbmc/bmcweb/blob/master/redfish-core/lib/log_services.hpp#L857

Without the additional dumpType added to the mapper query, the deleted
dumps would be returned in the query. Not sure if that's a mapper bug in
itself but specifying the dump type fixes the issue.

Tested:
- Verified with this change that the issue documented in the defect is
  no longer seen

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>