Default value for Name property
This commit implement changes to populate "Name" field for assembly,
Fabric adapter and Port schema.
Assuming that interface "xyz.openbmc_project.Inventory.Item" with
property "PrettyName" will not be populated for all the FRUs belonging
to these schemas, the name field is being by default filled with last
segment of Object path.
In case PrettyName is found, the value will be updated with the same.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>