SW508010 fix: Restrict VMI interface configuration to Manager role
Operator user privilege is able to configure VMI interface.
This is as per redfish specification. EthernetInterface operations
under the systems schema is allowed for Operator role.
But this is not a usecase for IBM's VMI interface.

This change is a downstream-only fix to change the patch operation
permission from "ConfigureComponents" to "ConfigureManager"

Signed-off-by: sunharis_in <sunharis@in.ibm.com>