Update the new URI for dump offload
This PR addresses the following defect:
https://w3.rchland.ibm.com/projects/bestquest/?verb=view&id=SW527615

There was a recent change in the AdditionalDataURI of
dumps. The management console reads the offload URI from
this field of the redfish response and then offloads the dump.
With this change, dumps were not getting offloaded.

This commit updates the offload URI in dump offload file.

Tested-By:

-X GET https://${bmc}/redfish/v1/Managers/bmc/LogServices/Dump/Entries/72/attachment > bmcdumpoffload
  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                 Dload  Upload   Total   Spent    Left  Speed
100 25348  100 25348    0     0   6418      0  0:00:03  0:00:03 --:--:--  6420

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>
Change-Id: I979c40e2cbe1458593d8b109083e326f7157f7d7