Fix DumpID in the Http Location header on dump creation
Earlier, when a dump is created, a task will be started
to monitor the dump creation and on successful dump creation,
the task returns the dump uri in the location field
of hhtp headers.

I found that the dump id is being appended twice, like
Resource_22 instead of Resource_2. This commit fixes the same.

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>
Change-Id: Iab8ac2708837ec3e54ed42fceabafe3e3fdd37dd