Fix unnecessary lock conflicts due to segment length validation
As per the current BMC algorithm, only first 2 bytes will
be checked irrespective of the segments.

For example:
say, there is a locks for the resource id:
288241371267989504 -  HEX 04 00 0A 00 00 00 00 00

Lock request on another resource with id:
288235873709850624 -  HEX 04 00 05 00 00 00 00 00

Here, when the second acquire lock request comes,
the bytes of each segment in the resource id will be
checked against the already available resorce id.

With the current implementation, only the first two
bytes of the resource ids are checked for every segment
flags and and hence, the second acquire lock request will fail.

This PR contains the fix, to check the bytes correctly
for each segment. Say, in the acquire lock request there
are 3 segment flags with segment length 1, 2, and 1, like
the following:

-d '{ "Request" :[ { "LockType":"Write",
                     "SegmentFlags": [
                         {"LockFlag":"DontLock","SegmentLength":1},
                         {"LockFlag":"DontLock","SegmentLength":2},
                         {"LockFlag":"LockAll","SegmentLength":1}
                     ],
                     "ResourceID": 288235873709850624 }
                 ]
    }'

the bytes comparison would be like the following:

Resource 1 id: 04 00 0A 00 00 00 00 00
Resource 2 id: 04 00 05 00 00 00 00 00

Byte comparison:
seg flag 1: (seg length 1)
4, 4

seg flag 2: (seg length 2)
0,0
10,5

seg flag 3: (seg length 1)
0,0

This PR addresses the following defect:
https://w3.rchland.ibm.com/projects/bestquest/?defect=SW528951

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>