Restrict VMI interface configuration to Manager role
The following PR has been lost during the migration of
node class to bmcweb routes:
https://github.ibm.com/openbmc/bmcweb/pull/163

Operator user privilege is able to configure VMI interface.
This is as per redfish specification. EthernetInterface operations
under the systems schema is allowed for Operator role.
But this is not a usecase for IBM's VMI interface.

This change is a downstream-only fix to change the patch operation
permission from "ConfigureComponents" to "ConfigureManager"

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>
Change-Id: I2287842c423e026ca56db0c78fc132be47d387df