Avoid patching of read-only bios attributes
Since the P10 system has BMC to manage the BIOS attributes,
and bmcweb does not know the attribute values of the BIOS,
it should allow the host to be able to set BIOS readOnly
attributes, while the redfish-doPATCH operation should not
be allowed.

With this change, when a redfish command to update a bios
attribute (which is read-only), is fired, bmcweb will send
the error back to the client.

Tested By:

Updation of read-only attr:

-X PATCH https://${bmc}/redfish/v1/Systems/system/Bios/Settings -d '{"Attributes":{"hb_debug_console":"Disabled"}}'
{
  "hb_debug_console@Message.ExtendedInfo": [
    {
      "@odata.type": "#Message.v1_1_1.Message",
      "Message": "The property hb_debug_console is a read only property and cannot be assigned a value.",
      "MessageArgs": [
        "hb_debug_console"
      ],
      "MessageId": "Base.1.8.1.PropertyNotWritable",
      "MessageSeverity": "Warning",
      "Resolution": "Remove the property from the request body and resubmit the request if the operation failed."
    }
  ]
}

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>