Accounts/service ACF properties added
For the 1120 rebase, merge all ACF and Service user commits into one.
This will ensure future cherry-picks of this downstream function are
easier to bring in. Having it spread across multiple commits makes the
process of bringing it back in during rebases unnecessarily complex.

The properties of the "service" user has been
    extended to include ["Oem"]["IBM"]["ACF"]

The patch and get methods to the route
    "/redfish/v1/AccountService/Accounts/service"
    now return the ACF properties to the request.

These properties include:
ACFFile, ExpirationDate, ACFInstalled,
WarningLongDatedExpiration, BMCCurrentDate

These were added in support for acf upload functionality.

A summary of these properties are as follows:
ACFFile : contents of the ACF file uploaded
ExpirationDate : date of expiration for the ACF file
ACFInstalled : ACF installed state
WarningLongDatedExpiration : set if expiration is
    > than 30 days from the BMC date
BMCCurrentDate : BMC current date to display if
    warning message is enabled

Like we have done in the past, ignore the error and just move on if
can't get non-critical property, in this case the acf date.

Signed-off-by: Noah Brewer <noah.t.brewer@gmail.com>
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>
Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>
Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Signed-off-by: Gunnar Mills <gmills@us.ibm.com>