Add support for SBE dump
SBE dump is basically the dump collected from
Self Boot Engine(SBE) when it encounters an error
and becomes unresponsive.

This commit contains changes that supports listing and
deletion of SBE dump. The SBE dumps will be listed under
/redfish/v1/Systems/system/LogServices/Dump along with
the system, resource hostboot and hardware dump entries
with the entry id prepended with "SBE_".

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>
Change-Id: I41eb4759cac88dd6eb7ae1ef3b4fff22f66a3975