Don't edit generated registries
As Teeks and I learned, if you run the parse registries script
downstream the outputted files don't match. This is becuase in
downstream we have a patch that edits the priviledge registry.

At the top of the file:
"// privilege_registry.hpp is generated.  Do not edit directly"

Respect that.

Downstream only since the following commit is downstream only:
https://github.com/ibm-openbmc/bmcweb/commit/cb2fecf012c7f20fa93b16bc799eb1a70f3c003b#diff-36d772115a7d28a9c1bbe1c848e05d4570875f056b876cec47513927088437f5

Tested: It builds. No Other testing.
Signed-off-by: Gunnar Mills <gmills@us.ibm.com>