Increase the dump create task timer
Currently, the task timer is set to 3 minutes to listen
for the dump completion status. But the bmc dumps are
taking a bit longer timer than 3 minutes (around 4 minutes)
during ipl and the task timer expires with the task state set
to cancelled. Hence, with this change, the task timer is doubled
to 6 minutes to avoid the task timer expiration.

Tested By:

Poweron the host
Initiate bmc dump
Get on the task for task status and the created dump URI
Task status - Completed

Fixes the defect:
https://w3.rchland.ibm.com/projects/bestquest/?defect=SW538120

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>