Remove stale lock table entries of invalid sessionid
Locks are cleaned up when there is a graceful logout /
session expiration due to timeout. But there are possibilities
of leaving stale lock entries in the BMC, if the session is
abruptly terminated.

This commit is to check for the existence of the session
whenever there is a lock conflict. If the session is not active,
remove the lock record considering that as a stale record and
return no error. If the session is active, return the lock
conflict to the client.

This pr fixes the defect at:
https://w3.rchland.ibm.com/projects/bestquest/?verb=view&id=SW538023&table=defect

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>
Change-Id: I1b97895cbdc4e602e00313bfd3b59e3b3445bd59