Add GeneratorId prop to redfish LogEntry (#136)
This commit adds a new oem property for user-trigerred
dumps, that will be used to indicate the unique id of
the user that triggered the dump creation. When multiple
clients manages the bmc, there was a requirement to know
that which client has initiated the dump before offloading/
performing any operations on that dump. Hence, this prop
tags the ip address of the user that has initiated the
dump creation.

There is a discussion at DMTF to add this property to
the redfish end-point on LogEntry schema. This can
be also used for audit purposes. Please see
https://redfishforum.com/thread/547/usage-originofcondition-parameter-logentry-schema

So right now, this property has been added to
oemlogentry schema that was added here:
https://github.com/ibm-openbmc/bmcweb/pull/121/files

The dbus interface change for the same is here:
https://github.com/ibm-openbmc/phosphor-dbus-interfaces/pull/26

Tested By:

Created a bmc dump using CollectDiagnosticData redfish api

GET on the created dump entry:
    {
      "@odata.id": "/redfish/v1/Managers/bmc/LogServices/Dump/Entries/43",
      "@odata.type": "#LogEntry.v1_8_0.LogEntry",
      "AdditionalDataSizeBytes": 108644,
      "AdditionalDataURI": "/redfish/v1/Managers/bmc/LogServices/Dump/Entries/43/attachment",
      "Created": "2021-11-02T10:49:37+00:00",
      "DiagnosticDataType": "Manager",
      "EntryType": "Event",
      "Id": "43",
      "Name": "BMC Dump Entry",
      "Oem": {
        "OpenBMC": {
          "@odata.type": "#OemLogEntry.v1_0_0.LogEntry",
          "GeneratorId": "::ffff:9.43.40.132"
        }
    }

Redfish Validator passed.
Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>
Change-Id: I050f2f2a9ff9a95ad6ac4fbc3a5a4283e26f64f7

Change-Id: I972102130c2393c3595adc8088792bc866abef83