1020 reduce error traces in bmcweb (#162)
* reduce error traces in priv and chassis code

These are not error paths so do not log them as errors

Tested:
- None other then CI. Changes are very simple.

* sensors: reduce error trace

This is a log that is only in our downstream code. Similar to an
upstream change in this area, this should not be an error trace.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>

* event:post-codes: log more info on error

It's not clear if this is really an error or not. It was seen a lot on
at least one system. For now, add the error code so we have a better
feel for what's failing.

This is downstream-only code so this is a downstream-only commit.

Change-Id: I65b99b466ba2d59c304cd80c88c2c81b940829b9
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>