HMC-BMC: SSLHandshakeException fix
During the system boot, huge request-reponse exchange between the
managemenet console and BMC intermittently causes the queue overflow
resulting in closing the tcp socket used for the communication.
The re-attempt requests and the other requests in flight gets SSL PeerShutdown
error and following requests in the queue will hit the SSL handshake
exception
This situation automatically recovers after the timer gets cancelled after
processing/failing the current requests in line. But by that time various
errors are hit at HMC and system will be moved to Incomplte/NoConnection/Error
state depending on the commands which randomly fail

This commit is to bump up the time queue size from 100 to 300, to make room
for the huge incoming traffix at bmcweb

Testedby:
  On a dual HMC setup:
  1. Several IPLs are performed and verified that no SSLHandshakeException seen
     at HMC
  2. Twenty-two partitions were created and some LPAR operations were done
  3. BMC reboot is tested when the system was at operating state

Signed-off-by: Sunitha Harish <sunharis@in.ibm.com>
Change-Id: Iad6c10705a02c612377347939c7c2bdbf54bb97f