Fix EnvironmentMetrics not showing
Noticed EnvironmentMetrics was not in the chassis resource.

EnvironmentMetrics is part of the new PowerSubsystem, ThermalSubsystem
so move it under the correct option.

Also moved PowerSubsystem to make it correct and fixed redfish.hpp.

https://gerrit.openbmc-project.xyz/c/openbmc/bmcweb/+/40326/26/redfish-core/lib/chassis.hpp
was correct upstream and left a comment on
https://gerrit.openbmc-project.xyz/c/openbmc/bmcweb/+/43170.

BMCWEB_ALLOW_DEPRECATED_POWER_THERMAL is disabled downstream while
BMCWEB_NEW_POWERSUBSYSTEM_THERMALSUBSYSTEM is enabled downstream.
https://github.com/ibm-openbmc/openbmc/blob/1020/meta-ibm/recipes-phosphor/interfaces/bmcweb_%25.bbappend#L4

Signed-off-by: Gunnar Mills <gmills@us.ibm.com>