OEM Battery Concurrent Maintenance For Everest System
The current impementation of this feature supports
the redfish OEM ReadyToRemove property for the TOD
battery on assembly.  The TOD battery is put in the
ReadyToRemove state by stopping the adcsensor application.
The reverse occurs when the adcsensor application
is restarted.  The adcsensor is stopped and started
by calling systemd.  Note that this implementation
only works if the adcsensor application handles one
ADC sensor.

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>