Push-style event: Delete subscriber after retry failures
Push style eventing implements sendEvent retries based on the
maximum retry limits set at the event service. When the max retry
is hit, it means that the subscriber is gone and bmc need not
attempt sending the event to that suscriber when next event occurs.
BMC should delete the subcription in this case.

Current implementation does not delete the subscription on hitting
maximum number of sendEvent retries. This overloads the http_client
at bmc to attempt a sendEvent to a non-existing or stale subscription.

This commit implements
1. Delete subscription after max retires
2. Store subscription id as a member of subscriber
3. Some trace enhancements at http client

Tested by:
 1. Subscribe to events and close the subscriber abruptly to see
    bmc retries and after all retry attempt failures, stops sending
    the events to that subscriber
 2. A bmc restart should cleanup these stale subscribers
 3. Tested sanity of the existing push-style eventing feature

Signed-off-by: Sunitha Harish <sunharis@in.ibm.com>
Change-Id: Idae9b4b4a8a4cc25239ea4637b7bde9f2e380b4f