OEM - Enclosure Firmware Version
Mex is updating via a lid in the host image, updated via PHYP.

PHYP sends an FRU record set PDR and sends the io enclosure version
data via the frurecord table, and pldm parses that and hosts the
Version Interface on the mex chassis object.

This implementation retrieves that mex chassis FirmwareVersion data,
which define via the manufacturer for the associated chassis, and
displayed on Redfish OEM schema for chassis.