Consume availability interface for cores
In the current state, the logic is bmcweb for displaying cores is:
It checks the Present property on core dbus object to update
the [Status][State] as Enabled/Absent.

In addition to Present property, we should also use the Available
property , if the Available is false, then set the [Status][State]
should be set as "UnavailableOffline".

Refer: https://github.com/ibm-openbmc/dev/issues/3481

Signed-off-by: George Liu <liuxiwei@inspur.com>