Reset Event retryPolicy on BMC reboot (#289)
When BMC reboots, the existing subscriptions get reloaded to the
memory. A http client connection object is created freshly. But
the retry policy values were not applied on that object and the
http client object uses the default values for them

This commit sets the retryPolicy and retryConfig parameters
after the BMC reboot for the persisted subscribers

Tested by:
  Check the retry policy and time intervals of a subscriber
after the BMC reset -or- bmcweb restart

Signed-off-by: sunharis <sunharis@in.ibm.com>
Change-Id: I572c1d1c7c56c609b392e28fb1e896dae5fe5b5f

Co-authored-by: sunharis <sunharis@in.ibm.com>