The following modifications are included:
1) Currently there is only one assembly that contains an OEM
entity. There must exist a method to identify the assembly that
contains the TOD battery other than the presence of an OEM property,
since more OEM properties could be added on other assemblies.
Previously the memberId associated with the battery OEM assembly
was used, but the memberId may change.  This modification identifies
the OEM battery assembly by its name on DBUS.

2)The memberId will not be required in the PATCH request when
only the Oem/OpenBMC/ReadyToRemove property for the battery
is contained in the PATCH request.

3) If the LocationIndicatorActive LED flag is included in the
PATCH request with the Oem/OpenBMC/ReadyToRemove property or
without the  Oem/OpenBMC/ReadyToRemove property, the memberId
will be required.

4) If the LocationIndicatorActive LED flag for any other assembly
is included in the PATCH request the associated memberId is required.

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>