Prevent error when occ-control server not available (#299)
The Power.Mode and Power.IdlePowerSaver data is only available when the
occ-control service is running. During a BMC reboot the service may
temporarly not be available so this data can not be provided.

Redfish clients would bet getting internalErrors when the power data is
not available. This change will allow the other system data to be returned.

Change-Id: I5ce64ebde9789b900daaca6e84554351b35dfbad
Signed-off-by: Chris Cain <cjcain@us.ibm.com>