bmc-state: remove legacy code
We pulled this commit into our downstream code:
https://gerrit.openbmc-project.xyz/c/openbmc/bmcweb/+/50318

But as noted in the last comment of that review, we should remove the
old code that tried to determine BMC state via systemd.

This shouldn't affect things too badly because of the static variable
and the fact this code only ever runs once. Otherwise, the code up
higher that calls getBMCState() will set this correctly.

Tested:
- None, seems obvious

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>