OEM Battery Concurrent Maintenance For Everest System (#224)
* OEM Battery Concurrent Maintenance For Everest System

The current impementation of this feature supports
the redfish OEM ReadyToRemove property for the TOD
battery on assembly.  The TOD battery is put in the
ReadyToRemove state by stopping the adcsensor application.
The reverse occurs when the adcsensor application
is restarted.  The adcsensor is stopped and started
by calling systemd.  Note that this implementation
only works if the adcsensor application handles one
ADC sensor.

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* Update for PR: #224

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* The current impementation of this feature supports the redfish OEM ReadyToRemove property
for the TOD battery on assembly. The TOD battery is put in the ReadyToRemove state by stopping
the adcsensor application. The reverse occurs when the adcsensor application is restarted.
The adcsensor is stopped and started by calling systemd. Note that this implementation only
works if the adcsensor application handles one ADC sensor.

The properties of an assemby have been extended to include ["Oem"]["OpenBMC"]["ReadyToRemove"]

The GEt and PATCH methods to the route "/redfish/v1/Chassis/chassis/Assembly" now handle the
property ["Oem"]["OpenBMC"]["ReadyToRemove"]

The following commands werE tested:
curl -k -X  GET https://${bmc}/redfish/v1/Chassis/chassis/Assembly

curl -k -X PATCH -d '{"Assemblies":[{"MemberId" : "1", "LocationIndicatorActive": true, "Oem":
{ "OpenBMC": { "ReadyToRemove": true }}}]}' https://${bmc}/redfish/v1/Chassis/chassis/Assembly

curl -k -X PATCH -d '{"Assemblies":[{"MemberId" : "1", "LocationIndicatorActive": false, "Oem":
{ "OpenBMC": { "ReadyToRemove": false }}}]}' https://${bmc}/redfish/v1/Chassis/chassis/Assembly

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* Update

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* Update For Review Comments

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* Update to PR for review comments.

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* Update for PR for review comments.

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* Update for PR comments:
Modifications were made to allow the LED LocationIndicatorActive
flag to be set independently of the OEM ReadyToRemove
property.

Currently either or both of the following sets of data
can be submitted in a BMCWEB PATCH request for the Battery
assembly.

1) "MemberId" : "1", "LocationIndicatorActive": true/false

2) "MemberId : "1", "Oem": { "OpenBMC": { "ReadyToRemove": true/false }}

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* Update for PR to delete debug message.

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* Modification to issue ERROR message if MemberID is
not set to "1", which is the battery assembly.

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* The following modifications are included:

1) Currently there is only one assembly that contains an OEM
entity. There must exist a method to identify the assembly that
contains the TOD battery other than the presence of an OEM property,
since more OEM properties could be added on other assemblies.
Previously the memberId associated with the battery OEM assembly
was used, but the memberId may change.  This modification identifies
the OEM battery assembly by its name on DBUS.

2)The memberId will not be required in the PATCH request when
only the Oem/OpenBMC/ReadyToRemove property for the battery
is contained in the PATCH request.

3) If the LocationIndicatorActive LED flag is included in the
PATCH request with the Oem/OpenBMC/ReadyToRemove property or
without the  Oem/OpenBMC/ReadyToRemove property, the memberId
will be required.

4) If the LocationIndicatorActive LED flag for any other assembly
is included in the PATCH request the associated memberId is required.

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* Update for PR comments.

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* Update for PR commants.

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* Update for PR comments.

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>

* Update for PR comments.

Signed-off-by: Tom Ippolito <thomas.ippolito@ibm.com>
Signed-off-by: Myung Bae <myungbae@us.ibm.com>