Define an OemPowerSupplyMetric schema
Create JSON and XML versions of the schema. JSON is more common, but XML
is what Refish uses. OpenBMC generally implements both.

Add the schema to the update script, and run the tool to get the schema into
the index.

InputPowerHistoryItem under IBM has the Date, Average, and Max arrays.

Change-Id: I7a67169dc6fdb17e422c4117f91ee3c71bf65818
Signed-off-by: Brandon Wyman <bjwyman@gmail.com>