Send event for VMI method change on ethernet intf
This change sends an event to the client when
there is a method change on the vmi interface from
static to DHCP and vice versa.

This is needed because, say, when the client changes method
from static to dhcp and if the dhcp is not connected, the
client wont receive any event from bmcweb, thus the cient thinks
that the method is static on that interface.

Tested By:
Connected HMC and verified whether the events are
being sent.

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>