PowerSubsystem: Ignore default MaxPowerCapValue
The Redfish Allocation properties in the PowerSubsystem are set based on
properties of the D-Bus xyz.openbmc_project.Control.Power.Cap interface.

The MaxPowerCapValue property on that interface defaults to a value of
UINT32_MAX.  Thus, that is the value before the actual maximum power cap
has been determined.

Update the PowerSubsystem implementation to treat the MaxPowerCapValue
value as invalid if it equals UINT32_MAX.  This avoids erroneous and
confusing Redfish output.

Tested:
* Test where MaxPowerCapValue is invalid
  * Value is 0
  * Value is UINT32_MAX
* Test where MaxPowerCapValue is valid

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>
Change-Id: I8c6fccb24c456310a13581968fdcfc2a53e94c97